chatFHIR-mcp-demo/mcp-servers/iris-mcp/requirements.txt
uv
mcpo
fastmcp
sqlalchemy
sqlalchemy-iris[intersystems]